backend/requirements.txt
pillow==8.2.0
python-multipart==0.0.5
aiofiles==0.7.0
boto3==1.17.98
loguru==0.5.3
pynamodb==5.1.0
uvicorn[standard]==0.14.0